re-enable multi-platform master images
This was inadvertently disabled by distributing the jobs on native
runners.

Configuring it for a single image is non-trivial:
https://docs.docker.com/build/ci/github-actions/multi-platform/#distribute-build-across-multiple-runners

And for a matrix of images more difficult still, without duplicating the
workflow.

Revert to QEMU and using --platforms to get multi-platform.

This could likely be speeded up by using cross-compilation in the Dockerfiles if necessary.